COMMERCE-8276 dispatch-test rename messaging/test to messaging
